Add support for the blocking HTTP client being optional (based on Cargo features).
