+++ Cargo.toml
@@ -26,7 +26,7 @@ optional = true
 version = "2.33"
 
 [dependencies.reqwest]
-features = ["blocking", "json"]
+features = ["json"]
 version = "0.11"
 
 [dependencies.serde]
@@ -34,4 +34,6 @@ features = ["derive"]
 version = "1.0.93"
 
 [features]
+blocking = ["reqwest/blocking"]
 cli = ["clap"]
+default = ["blocking"]
